Fix enough NOLF1 headers working to compile one source file.

+++ NOLF/ClientShellDLL/BitmapCtrl.h
@@ -28,8 +28,8 @@ public:
 	virtual void	Render ( HSURFACE hDestSurf );
 
 	// Width/Height calculations
-	virtual int		GetWidth ( );
-	virtual int		GetHeight ( );
+	virtual uint16		GetWidth ( );
+	virtual uint16		GetHeight ( );
 
     virtual LTBOOL   OnEnter();
     virtual LTBOOL   OnLButtonUp (int x, int y) {return OnEnter();}
